Add in analog pedal support Add analog pedal value allocation Added analog and digital pedal value readouts to debug window Add pedal outputs to MIDI output MIDI events moved such that they only get created when necessary Miscellaneous: - Changed pedalAnalog to be a byte - Changed internal octave value to be a byte

+++ InOut/InputState.cs
@@ -59,7 +59,7 @@ namespace InOut
 		/// <remarks>
 		/// Unused because I don't know how this works, or if this is even a functionality lol, don't have an analog pedal to test with.
 		/// </remarks>
-		public int pedalAnalog;	
+		public byte pedalAnalog;	
 
 		/// <summary>
 		/// The face buttons.
